requirements.txt
torch
nougat-ocr -q
gradio -U -q